# Conflicts:
#	Data/Sys/GameSettings/FA6.ini
#	Data/Sys/GameSettings/FA7.ini
#	Data/Sys/GameSettings/FA8.ini
#	Data/Sys/GameSettings/FA9.ini
#	Data/Sys/GameSettings/FAA.ini
#	Data/Sys/GameSettings/FAB.ini
#	Data/Sys/GameSettings/FACE01.ini
#	Data/Sys/GameSettings/FACP01.ini
#	Data/Sys/GameSettings/FAFE01.ini
#	Data/Sys/GameSettings/FAG.ini
#	Data/Sys/GameSettings/FAH.ini
#	Data/Sys/GameSettings/FAL.ini
#	Data/Sys/GameSettings/FAME01.ini
#	Data/Sys/GameSettings/FAN.ini
#	Data/Sys/GameSettings/FAO.ini
#	Data/Sys/GameSettings/FAP.ini
#	Data/Sys/GameSettings/FAQ.ini
#	Data/Sys/GameSettings/FARE01.ini
#	Data/Sys/GameSettings/FAS.ini
#	Data/Sys/GameSettings/FAT.ini
#	Data/Sys/GameSettings/FAV.ini
#	Data/Sys/GameSettings/FAW.ini
#	Data/Sys/GameSettings/FAX.ini
#	Data/Sys/GameSettings/FB4.ini
#	Data/Sys/GameSettings/FB5.ini
#	Data/Sys/GameSettings/FBB.ini
#	Data/Sys/GameSettings/FBC.ini
#	Data/Sys/GameSettings/FBD.ini
#	Data/Sys/GameSettings/FBE.ini
#	Data/Sys/GameSettings/FBH.ini
#	Data/Sys/GameSettings/FBI.ini
#	Data/Sys/GameSettings/FBJ.ini
#	Data/Sys/GameSettings/FBL.ini
#	Data/Sys/GameSettings/FBN.ini
#	Data/Sys/GameSettings/FBR.ini
#	Data/Sys/GameSettings/FBS.ini
#	Data/Sys/GameSettings/FBU.ini
#	Data/Sys/GameSettings/FBY.ini
#	Data/Sys/GameSettings/FBZ.ini
#	Data/Sys/GameSettings/FC3.ini
#	Data/Sys/GameSettings/FC6.ini
#	Data/Sys/GameSettings/FC7.ini
#	Data/Sys/GameSettings/FC8.ini
#	Data/Sys/GameSettings/FCA.ini
#	Data/Sys/GameSettings/FCP.ini
#	Data/Sys/GameSettings/FCQ.ini
#	Data/Sys/GameSettings/FCR.ini
#	Data/Sys/GameSettings/FCS.ini
#	Data/Sys/GameSettings/FCT.ini
#	Data/Sys/GameSettings/FCU.ini
#	Data/Sys/GameSettings/FCV.ini
#	Data/Sys/GameSettings/FCW.ini
#	Data/Sys/GameSettings/FCY.ini
#	Data/Sys/GameSettings/FCZ.ini
#	Data/Sys/GameSettings/FD2.ini
#	Data/Sys/GameSettings/FD7.ini
#	Data/Sys/GameSettings/FDA.ini
#	Data/Sys/GameSettings/FDF.ini
#	Data/Sys/GameSettings/FDL.ini
#	Data/Sys/GameSettings/FDN.ini
#	Data/Sys/GameSettings/FDO.ini
#	Data/Sys/GameSettings/FDP.ini
#	Data/Sys/GameSettings/FDQ.ini
#	Data/Sys/GameSettings/FDT.ini
#	Data/Sys/GameSettings/FDU.ini
#	Data/Sys/GameSettings/FDV.ini
#	Data/Sys/GameSettings/FEC.ini
#	Data/Sys/GameSettings/FED.ini
#	Data/Sys/GameSettings/FEI.ini
#	Data/Sys/GameSettings/FEM.ini
#	Data/Sys/GameSettings/FEN.ini
#	Data/Sys/GameSettings/FEQ.ini
#	Data/Sys/GameSettings/FER.ini
#	Data/Sys/GameSettings/FES.ini
#	Data/Sys/GameSettings/FEU.ini
#	Data/Sys/GameSettings/FFA.ini
#	Data/Sys/GameSettings/FFD.ini
#	Data/Sys/GameSettings/FFE.ini
#	Data/Sys/GameSettings/FFL.ini
#	Data/Sys/GameSettings/FFM.ini
#	Data/Sys/GameSettings/FFN.ini
#	Data/Sys/GameSettings/FFQ.ini
#	Data/Sys/GameSettings/G8M.ini
#	Data/Sys/GameSettings/GAF.ini
#	Data/Sys/GameSettings/GAL.ini
#	Data/Sys/GameSettings/GL5.ini
#	Data/Sys/GameSettings/GL7.ini
#	Data/Sys/GameSettings/GLM.ini
#	Data/Sys/GameSettings/GM4.ini
#	Data/Sys/GameSettings/GMS.ini
#	Data/Sys/GameSettings/GYQ.ini
#	Data/Sys/GameSettings/GZL.ini
#	Data/Sys/GameSettings/JAA.ini
#	Data/Sys/GameSettings/JAD.ini
#	Data/Sys/GameSettings/JAI.ini
#	Data/Sys/GameSettings/JAV.ini
#	Data/Sys/GameSettings/JCB.ini
#	Data/Sys/GameSettings/JDA.ini
#	Data/Sys/GameSettings/LAE.ini
#	Data/Sys/GameSettings/LAN.ini
#	Data/Sys/GameSettings/LAQ.ini
#	Data/Sys/GameSettings/MAE.ini
#	Data/Sys/GameSettings/MAF.ini
#	Data/Sys/GameSettings/MAH.ini
#	Data/Sys/GameSettings/MAJ.ini
#	Data/Sys/GameSettings/MAX.ini
#	Data/Sys/GameSettings/MAZ.ini
#	Data/Sys/GameSettings/MBD.ini
#	Data/Sys/GameSettings/MBO.ini
#	Data/Sys/GameSettings/NAB.ini
#	Data/Sys/GameSettings/NAL.ini
#	Data/Sys/GameSettings/NAN.ini
#	Data/Sys/GameSettings/PAC.ini
#	Data/Sys/GameSettings/PZLE01.ini
#	Data/Sys/GameSettings/R8P.ini
#	Data/Sys/GameSettings/RMC.ini
#	Data/Sys/GameSettings/RSB.ini
#	Data/Sys/GameSettings/RUU.ini
#	Data/Sys/GameSettings/SB4.ini
#	Data/Sys/GameSettings/SF8.ini
#	Data/Sys/GameSettings/SMN.ini
#	Readme.md
#	Source/Core/AudioCommon/Mixer.cpp
#	Source/Core/Common/GL/GLInterface/WGL.h
#	Source/Core/Common/GL/GLInterfaceBase.h
#	Source/Core/Common/IniFile.h
#	Source/Core/Common/MathUtil.cpp
#	Source/Core/Common/MathUtil.h
#	Source/Core/Common/StringUtil.cpp
#	Source/Core/Common/StringUtil.h
#	Source/Core/Core/Boot/Boot.cpp
#	Source/Core/Core/Boot/Boot_BS2Emu.cpp
#	Source/Core/Core/ConfigManager.cpp
#	Source/Core/Core/Core.cpp
#	Source/Core/Core/Core.h
#	Source/Core/Core/Core.vcxproj
#	Source/Core/Core/Core.vcxproj.filters
#	Source/Core/Core/HW/Memmap.cpp
#	Source/Core/Core/HW/SystemTimers.cpp
#	Source/Core/Core/HW/VideoInterface.h
#	Source/Core/Core/HW/WiimoteEmu/Attachment/Classic.cpp
#	Source/Core/Core/HotkeyManager.cpp
#	Source/Core/Core/HotkeyManager.h
#	Source/Core/Core/PowerPC/Interpreter/Interpreter_Branch.cpp
#	Source/Core/Core/PowerPC/JitCommon/JitBackpatch.cpp
#	Source/Core/DiscIO/DiscScrubber.cpp
#	Source/Core/DolphinQt/GameList/GameFile.cpp
#	Source/Core/DolphinQt/GameList/GameFile.h
#	Source/Core/DolphinQt2/DolphinQt2.vcxproj
#	Source/Core/DolphinWX/AboutDolphin.cpp
#	Source/Core/DolphinWX/Config/GeneralConfigPane.cpp
#	Source/Core/DolphinWX/Config/GeneralConfigPane.h
#	Source/Core/DolphinWX/DolphinWX.vcxproj
#	Source/Core/DolphinWX/DolphinWX.vcxproj.filters
#	Source/Core/DolphinWX/Frame.cpp
#	Source/Core/DolphinWX/GameListCtrl.cpp
#	Source/Core/DolphinWX/ISOProperties.cpp
#	Source/Core/DolphinWX/MainNoGUI.cpp
#	Source/Core/DolphinWX/VideoConfigDiag.cpp
#	Source/Core/DolphinWX/VideoConfigDiag.h
#	Source/Core/InputCommon/ControllerInterface/ControllerInterface.cpp
#	Source/Core/InputCommon/InputCommon.vcxproj
#	Source/Core/InputCommon/InputCommon.vcxproj.filters
#	Source/Core/VideoBackends/D3D/D3DBase.cpp
#	Source/Core/VideoBackends/D3D/D3DShader.cpp
#	Source/Core/VideoBackends/D3D/FramebufferManager.cpp
#	Source/Core/VideoBackends/D3D/PixelShaderCache.cpp
#	Source/Core/VideoBackends/D3D/PixelShaderCache.h
#	Source/Core/VideoBackends/D3D/Render.cpp
#	Source/Core/VideoBackends/D3D/main.cpp
#	Source/Core/VideoBackends/OGL/FramebufferManager.cpp
#	Source/Core/VideoBackends/OGL/FramebufferManager.h
#	Source/Core/VideoBackends/OGL/ProgramShaderCache.h
#	Source/Core/VideoBackends/OGL/Render.cpp
#	Source/Core/VideoBackends/OGL/Render.h
#	Source/Core/VideoBackends/OGL/main.cpp
#	Source/Core/VideoBackends/Software/SWRenderer.cpp
#	Source/Core/VideoBackends/Software/SWmain.cpp
#	Source/Core/VideoBackends/Software/VideoBackend.h
#	Source/Core/VideoCommon/BPStructs.cpp
#	Source/Core/VideoCommon/CommandProcessor.cpp
#	Source/Core/VideoCommon/Fifo.cpp
#	Source/Core/VideoCommon/GeometryShaderGen.cpp
#	Source/Core/VideoCommon/GeometryShaderGen.h
#	Source/Core/VideoCommon/GeometryShaderManager.cpp
#	Source/Core/VideoCommon/MainBase.cpp
#	Source/Core/VideoCommon/OpcodeDecoding.cpp
#	Source/Core/VideoCommon/OpcodeDecoding.h
#	Source/Core/VideoCommon/PixelShaderGen.cpp
#	Source/Core/VideoCommon/RenderBase.cpp
#	Source/Core/VideoCommon/ShaderGenCommon.h
#	Source/Core/VideoCommon/TextureCacheBase.cpp
#	Source/Core/VideoCommon/TextureCacheBase.h
#	Source/Core/VideoCommon/TextureConversionShader.cpp
#	Source/Core/VideoCommon/VertexLoaderManager.cpp
#	Source/Core/VideoCommon/VertexManagerBase.cpp
#	Source/Core/VideoCommon/VertexShaderGen.cpp
#	Source/Core/VideoCommon/VertexShaderManager.cpp
#	Source/Core/VideoCommon/VideoBackendBase.h
#	Source/Core/VideoCommon/VideoCommon.vcxproj
#	Source/Core/VideoCommon/VideoCommon.vcxproj.filters
#	Source/Core/VideoCommon/VideoConfig.cpp
#	Source/Core/VideoCommon/VideoConfig.h
#	Source/Core/VideoCommon/XFStructs.cpp
#	Source/PCH/pch.h
#	Source/UnitTests/UnitTests.vcxproj
#	Source/dolphin-emu.sln

+++ Source/Android/app/src/main/java/org/dolphinemu/dolphinemu/services/AssetCopyService.java
@@ -10,10 +10,9 @@ import android.app.IntentService;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
-import android.util.Log;
 
 import org.dolphinemu.dolphinemu.NativeLibrary;
-import org.dolphinemu.dolphinemu.utils.UserPreferences;
+import org.dolphinemu.dolphinemu.utils.Log;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -27,8 +26,6 @@ import java.io.OutputStream;
  */
 public final class AssetCopyService extends IntentService
 {
-	private static final String TAG = "DolphinEmulator";
-
 	public AssetCopyService()
 	{
 		// Superclass constructor is called to name the thread on which this service executes.
@@ -40,32 +37,17 @@ public final class AssetCopyService extends IntentService
 	{
 		String BaseDir = NativeLibrary.GetUserDirectory();
 		String ConfigDir = BaseDir + File.separator + "Config";
-		String GCDir = BaseDir + File.separator + "GC";
 
 		// Copy assets if needed
-		File file = new File(GCDir + File.separator + "font_sjis.bin");
-		if(!file.exists())
-		{
-			NativeLibrary.CreateUserFolders();
-			copyAsset("dsp_coef.bin", GCDir + File.separator + "dsp_coef.bin");
-			copyAsset("dsp_rom.bin", GCDir + File.separator + "dsp_rom.bin");
-			copyAsset("font_ansi.bin", GCDir + File.separator + "font_ansi.bin");
-			copyAsset("font_sjis.bin", GCDir + File.separator + "font_sjis.bin");
-			copyAssetFolder("Shaders", BaseDir + File.separator + "Shaders");
-		}
-		else
-		{
-			Log.v(TAG, "Skipping asset copy operation.");
-		}
+		NativeLibrary.CreateUserFolders();
+		copyAssetFolder("GC", BaseDir + File.separator + "GC", false);
+		copyAssetFolder("Shaders", BaseDir + File.separator + "Shaders", false);
+		copyAssetFolder("Wii", BaseDir + File.separator + "Wii", false);
 
 		// Always copy over the GCPad config in case of change or corruption.
 		// Not a user configurable file.
-		copyAsset("GCPadNew.ini", ConfigDir + File.separator + "GCPadNew.ini");
-		copyAsset("WiimoteNew.ini", ConfigDir + File.separator + "WiimoteNew.ini");
-
-		// Load the configuration keys set in the Dolphin ini and gfx ini files
-		// into the application's shared preferences.
-		UserPreferences.LoadIniToPrefs(this);
+		copyAsset("GCPadNew.ini", ConfigDir + File.separator + "GCPadNew.ini", true);
+		copyAsset("WiimoteNew.ini", ConfigDir + File.separator + "WiimoteNew.ini", true);
 
 		// Record the fact that we've done this before, so we don't do it on every launch.
 		SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(this);
@@ -75,40 +57,45 @@ public final class AssetCopyService extends IntentService
 		editor.commit();
 	}
 
-	private void copyAsset(String asset, String output)
+	private void copyAsset(String asset, String output, Boolean overwrite)
 	{
-		Log.v(TAG, "Copying " + asset + " to " + output);
+		Log.verbose("[AssetCopyService] Copying File " + asset + " to " + output);
 		InputStream in = null;
 		OutputStream out = null;
 
 		try
 		{
-			in = getAssets().open(asset);
-			out = new FileOutputStream(output);
-			copyFile(in, out);
-			in.close();
-			out.close();
+			File file = new File(output);
+			if(!file.exists() || overwrite)
+			{
+				in = getAssets().open(asset);
+				out = new FileOutputStream(output);
+				copyFile(in, out);
+				in.close();
+				out.close();
+			}
 		}
 		catch (IOException e)
 		{
-			Log.e(TAG, "Failed to copy asset file: " + asset, e);
+			Log.error("[AssetCopyService] Failed to copy asset file: " + asset + e.getMessage());
 		}
 	}
 
-	private void copyAssetFolder(String assetFolder, String outputFolder)
+	private void copyAssetFolder(String assetFolder, String outputFolder, Boolean overwrite)
 	{
-		Log.v(TAG, "Copying " + assetFolder + " to " + outputFolder);
+		Log.verbose("[AssetCopyService] Copying Folder " + assetFolder + " to " + outputFolder);
 
 		try
 		{
 			for (String file : getAssets().list(assetFolder))
 			{
-				copyAsset(assetFolder + File.separator + file, outputFolder + File.separator + file);
+				copyAssetFolder(assetFolder + File.separator + file, outputFolder + File.separator + file, overwrite);
+				copyAsset(assetFolder + File.separator + file, outputFolder + File.separator + file, overwrite);
 			}
 		}
 		catch (IOException e)
 		{
-			Log.e(TAG, "Failed to copy asset folder: " + assetFolder, e);
+			Log.error("[AssetCopyService] Failed to copy asset folder: " + assetFolder + e.getMessage());
 		}
 	}
